requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
numpy<2.0.0
opencv-python-headless==4.8.1.78
mysql-connector-python==8.1.0
